requirements/base.txt
alembic
dynaconf
sqlalchemy
flask-migrate
marshmallow
marshmallow-enum
python-dotenv
psycopg2
flask-restful
flask==2.0.1
webargs
requests
faktory
apscheduler
grpcio-tools
zstd
gunicorn
python-snappy
flask_cors